refactor: Modify const values and separate related functionality
